@@ -1,56 +1,72 @@
 {
     "type": "project",
     "license": "proprietary",
+    "minimum-stability": "dev",
+    "prefer-stable": true,
     "require": {
-        "php": "^7.1.3",
+        "php": ">=7.2.5",
         "ext-ctype": "*",
         "ext-iconv": "*",
-        "captcha-com/symfony-captcha-bundle": "4.*",
-        "cocur/slugify": "^3.1",
-        "friendsofsymfony/jsrouting-bundle": "^2.3",
-        "knplabs/knp-paginator-bundle": "^4.0",
-        "oneup/uploader-bundle": "^2.1",
-        "predis/predis": "^1.1",
+        "composer/package-versions-deprecated": "1.11.99.1",
+        "doctrine/annotations": "^1.0",
+        "doctrine/doctrine-bundle": "^2.3",
+        "doctrine/doctrine-migrations-bundle": "^3.1",
+        "doctrine/orm": "^2.8",
+        "knplabs/knp-paginator-bundle": "^5.4",
+        "phpdocumentor/reflection-docblock": "^5.2",
         "sensio/framework-extra-bundle": "^5.1",
-        "snc/redis-bundle": "3.0.0-RC2",
-        "stof/doctrine-extensions-bundle": "^1.3",
-        "symfony/asset": "4.2.*",
-        "symfony/console": "4.2.*",
-        "symfony/dotenv": "4.2.*",
-        "symfony/expression-language": "4.2.*",
-        "symfony/flex": "^1.1",
-        "symfony/form": "4.2.*",
-        "symfony/framework-bundle": "4.2.*",
-        "symfony/intl": "4.2.*",
-        "symfony/lock": "4.2.*",
+        "stof/doctrine-extensions-bundle": "^1.6",
+        "symfony/asset": "5.2.*",
+        "symfony/console": "5.2.*",
+        "symfony/dotenv": "5.2.*",
+        "symfony/expression-language": "5.2.*",
+        "symfony/flex": "^1.3.1",
+        "symfony/form": "5.2.*",
+        "symfony/framework-bundle": "5.2.*",
+        "symfony/http-client": "5.2.*",
+        "symfony/intl": "5.2.*",
+        "symfony/lock": "5.2.*",
+        "symfony/mailer": "5.2.*",
+        "symfony/mime": "5.2.*",
         "symfony/monolog-bundle": "^3.1",
-        "symfony/orm-pack": "*",
-        "symfony/process": "4.2.*",
-        "symfony/security-bundle": "4.2.*",
-        "symfony/serializer-pack": "*",
-        "symfony/swiftmailer-bundle": "^3.1",
-        "symfony/translation": "4.2.*",
-        "symfony/twig-bundle": "4.2.*",
-        "symfony/validator": "4.2.*",
-        "symfony/web-link": "4.2.*",
-        "symfony/webpack-encore-bundle": "^1.7",
-        "symfony/yaml": "4.2.*",
-        "tetranz/select2entity-bundle": "^2.9",
+        "symfony/notifier": "5.2.*",
+        "symfony/process": "5.2.*",
+        "symfony/property-access": "5.2.*",
+        "symfony/property-info": "5.2.*",
+        "symfony/proxy-manager-bridge": "5.2.*",
+        "symfony/security-bundle": "5.2.*",
+        "symfony/serializer": "5.2.*",
+        "symfony/string": "5.2.*",
+        "symfony/swiftmailer-bundle": "^3.5",
+        "symfony/translation": "5.2.*",
+        "symfony/twig-bundle": "^5.2",
+        "symfony/validator": "5.2.*",
+        "symfony/web-link": "5.2.*",
+        "symfony/webpack-encore-bundle": "^1.11",
+        "symfony/yaml": "5.2.*",
+        "tetranz/select2entity-bundle": "^3.1",
         "twig/extensions": "^1.5",
-        "willdurand/js-translation-bundle": "^2.6"
+        "twig/extra-bundle": "^2.12|^3.0",
+        "twig/twig": "^2.12|^3.0",
+        "willdurand/js-translation-bundle": "^4.0"
     },
     "require-dev": {
-        "symfony/debug-pack": "*",
+        "rector/rector": "^0.10.17",
+        "symfony/browser-kit": "^5.2",
+        "symfony/css-selector": "^5.2",
+        "symfony/debug-bundle": "^5.2",
         "symfony/maker-bundle": "^1.0",
-        "symfony/profiler-pack": "*",
-        "symfony/test-pack": "*",
-        "symfony/web-server-bundle": "4.2.*"
+        "symfony/phpunit-bridge": "^5.2",
+        "symfony/stopwatch": "^5.2",
+        "symfony/var-dumper": "^5.2",
+        "symfony/web-profiler-bundle": "^5.2"
     },
     "config": {
+        "optimize-autoloader": true,
         "preferred-install": {
             "*": "dist"
         },
-        "sort-packages": true       
+        "sort-packages": true
     },
     "autoload": {
         "psr-4": {
