@@ -14,26 +14,26 @@ use App\Entity\Mailaddr;
  */
 class Wblist
 {
-  const WBLIST_PRIORITY_DOMAIN = 0;
-  const WBLIST_PRIORITY_GROUP = 50;
-  const WBLIST_PRIORITY_USER = 100;
-  const WBLIST_PRIORITY_GROUP_OVERRIDE = 200;
+    const WBLIST_PRIORITY_DOMAIN = 0;
+    const WBLIST_PRIORITY_GROUP = 50;
+    const WBLIST_PRIORITY_USER = 100;
+    const WBLIST_PRIORITY_GROUP_OVERRIDE = 200;
 
-  
+
     /**
     * @ORM\ManyToOne(targetEntity="App\Entity\User")
     * @ORM\JoinColumn(name="rid", nullable=true, onDelete="CASCADE")
     * @ORM\Id
     */
-    private $rid;       
+    private $rid;
 
     /**
     * @ORM\ManyToOne(targetEntity="App\Entity\Mailaddr")
     * @ORM\JoinColumn(name="sid", nullable=true)
     * @ORM\Id
     */
-    private $sid; 
-    
+    private $sid;
+
     /**
      * @var string
      *
@@ -54,7 +54,7 @@ class Wblist
      * @ORM\Column(name="type", type="integer", nullable=true)
      */
     private $type;
-    
+
     /**
     * @ORM\ManyToOne(targetEntity="App\Entity\Groups")
     * @ORM\JoinColumn(name="group_id", nullable=true, onDelete="CASCADE")
@@ -64,9 +64,9 @@ class Wblist
     /**
      * @ORM\Column(type="integer", nullable=true)
      */
-    private $priority;    
-    
-    public function __construct(User $user,Mailaddr $mailaddr)
+    private $priority;
+
+    public function __construct(User $user, Mailaddr $mailaddr)
     {
         $this->rid = $user;
         $this->sid = $mailaddr;
@@ -156,6 +156,4 @@ class Wblist
 
         return $this;
     }
-
-
 }
