@@ -4,7 +4,7 @@ namespace App\Repository;
 
 use App\Entity\Policy;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Symfony\Bridge\Doctrine\RegistryInterface;
+use Doctrine\Persistence\ManagerRegistry;
 use Cocur\Slugify\Slugify;
 
 /**
@@ -13,11 +13,11 @@ use Cocur\Slugify\Slugify;
  * @method Policy[]    findAll()
  * @method Policy[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
  */
-class PolicyRepository extends ServiceEntityRepository {
+class PolicyRepository extends ServiceEntityRepository
+{
 
-    public function __construct(RegistryInterface $registry) {
+    public function __construct(ManagerRegistry $registry)
+    {
         parent::__construct($registry, Policy::class);
     }
-
-    
 }
