@@ -1,24 +1,27 @@
 <?php
+
 namespace App\Service;
+
 use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
 
 class CryptEncrypt
 {
 
   //  private $params;
-    
+
     public function __construct(ParameterBagInterface $params)
     {
         $this->params = $params;
     }
-    
+
 
     /**
      * Encrypt URL for password reinitialization
      * @param type $url
      * @return type
      */
-    public function encryptUrl($url) {
+    public function encryptUrl($url)
+    {
         $salt = $this->params->get('salt');
         $iv = $this->params->get('iv');
         $crypt = openssl_encrypt($url, 'seed', $salt, 0, $iv);
@@ -34,13 +37,13 @@ class CryptEncrypt
      * @param type $token
      * @return type
      */
-    public function decryptUrl($token) {
+    public function decryptUrl($token)
+    {
         $salt = $this->params->get('salt');
         $iv = $this->params->get('iv');
         $cryptdate = substr(base64_decode($token), 0, 10);
         $crypt = substr_replace(base64_decode($token), '', 0, 10);
         $decrypted = openssl_decrypt($crypt, 'seed', $salt, 0, $iv);
-        return array($cryptdate, $decrypted);
+        return [$cryptdate, $decrypted];
     }
-    
 }
