@@ -10,16 +10,15 @@ use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorage;
 class UploaderNamer implements NamerInterface
 {
     private $tokenStorage;
-    
+
     public function __construct(TokenStorage $tokenStorage)
     {
         $this->tokenStorage = $tokenStorage;
     }
-    
+
     public function name(FileInterface $file)
     {
 
-      return 'logo.' . $file->getExtension();
-
+        return 'logo.' . $file->getExtension();
     }
 }
